+++ dataquality-statistics/src/test/java/org/talend/dataquality/statistics/datetime/SystemDateTimePatternManagerTest.java
@@ -14,6 +14,8 @@ package org.talend.dataquality.statistics.datetime;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assume.assumeTrue;
 
@@ -102,18 +104,21 @@ public class SystemDateTimePatternManagerTest {
     }
 
     @Test
-    public void testgetDateTimeFormatterByPattern() {
+    public void testGetDateTimeFormatterByPattern1() {
         DateTimeFormatter dateTimeFormatterByPattern =
                 SystemDateTimePatternManager.getDateTimeFormatterByPattern("dd/MM/yyyy", Locale.ENGLISH);
-        assertFalse(dateTimeFormatterByPattern == null);
-        assertTrue(dateTimeFormatterByPattern.getResolverStyle() == ResolverStyle.STRICT);
+        assertNotNull(dateTimeFormatterByPattern);
+        assertEquals(dateTimeFormatterByPattern.getResolverStyle(), ResolverStyle.STRICT);
         assertEquals("17/08/2015", dateTimeFormatterByPattern.format(LocalDate.of(2015, 8, 17)));
-        dateTimeFormatterByPattern =
+    }
+
+    @Test
+    public void testGetDateTimeFormatterByPattern2() {
+        DateTimeFormatter dateTimeFormatterByPattern =
                 SystemDateTimePatternManager.getDateTimeFormatterByPattern("yyyy-MM-dd G", Locale.US);
-        assertFalse(dateTimeFormatterByPattern == null);
+        assertNotNull(dateTimeFormatterByPattern);
         assertEquals("2015-08-17 AD", dateTimeFormatterByPattern.format(LocalDate.of(2015, 8, 17)));
-        dateTimeFormatterByPattern = SystemDateTimePatternManager.getDateTimeFormatterByPattern("yyyy-MM-dd G", null);
-        assertTrue(dateTimeFormatterByPattern == null);
+        assertNull(SystemDateTimePatternManager.getDateTimeFormatterByPattern("yyyy-MM-dd G", null));
     }
 
 }